Commit d1566ea9fad98129d40f46c7064b9af1dc8cd312 exposed a broken test: The timeuntil filter never raised an error in the first place.
I removed the test completely (at least as long as we're only relying on Django behavior to raise the errors for us). But as I noted elsewhere, maybe Coffin shouldn't be in the business of trying to enforce additional strict error checking on top of Django's builtin stuff in the first place.