Interop can now preprocess Jinja 2 "Undefined" values to make them usable with filters written for Django.
This breaks a test for our Jinja2-specific date filter. The core problem is that we don't have a way to register a filter while saying "this was written for Jinja" or "this was written for Django". Thus the interop layer is always applied, and no filter registered via Coffin can every get to see a "Undefined". We need to make some slightly for fundamental changes to fix this.